refactor(seo): /lpを削除しLPコンテンツを/に統合する
