+++ src/app/sitemap.ts
@@ -4,8 +4,8 @@ import { env } from "@/lib/env";
 /**
  * /sitemap.xml を動的生成する。
  *
- * 出力するURL: / (トップ)、/lp、/terms、/privacy、/legal
- * - /lp を最高優先度 (priority: 1.0) に設定
+ * 出力するURL: / (トップ)、/terms、/privacy、/legal
+ * - / を最高優先度 (priority: 1.0) に設定
  * - 主要ページは changeFrequency: "monthly"
  */
 export default function sitemap(): MetadataRoute.Sitemap {
@@ -14,12 +14,6 @@ export default function sitemap(): MetadataRoute.Sitemap {
   return [
     {
       url: `${baseUrl}/`,
-      lastModified: new Date(),
-      changeFrequency: "monthly",
-      priority: 0.9,
-    },
-    {
-      url: `${baseUrl}/lp`,
       lastModified: new Date(),
       changeFrequency: "monthly",
       priority: 1.0,
